requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.3
blis==0.7.11
build==1.2.1
catalogue==2.0.10
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cryptography==42.0.7
cymem==2.0.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
orjson==3.10.3
packaging==24.1
passlib==1.7.4
pip-tools==7.4.1
preshed==3.0.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.4
pydantic-extra-types==2.7.0
pydantic-settings==2.0.3
pydantic_core==2.18.4
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
thinc==8.2.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.0
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
